buck2: rewrite uv_tool and inject_crc
for some reason, genrules can't be used in other rules. This makes my
original implementation of uv_tool kinda sucky because it means
you can't write a rule that leverages uv as part of it.

Here, I've rewritten uv_tool in such a way that it can support both:
 1. uv_genrule (no change to the customer-facing aspect here)
 2. use within rules, specifically in `ctx.actions.run`

Then, I rewrote `inject_crc` as a rule leveraging the functionality
described in 2. above